+++ gateway/src/main/java/org/acumos/federation/gateway/service/CatalogServiceConfiguration.java
@@ -51,6 +51,7 @@ public class CatalogServiceConfiguration {
 	private static final EELFLoggerDelegate log = EELFLoggerDelegate.getLogger(MethodHandles.lookup().lookupClass());
 
 	private Map<String, Object>	solutionsSelector;
+	private Map<String, Object>	solutionsSelectorDefaults;
 	private Map<String, Object>	solutionRevisionsSelector;
 
 
@@ -64,13 +65,16 @@ public class CatalogServiceConfiguration {
 	 * the corresponding field is no longer maintained in the backend. It ca always be re-introduced through configuration.
 	 */
 	private void reset() {
+		this.solutionsSelectorDefaults = new HashMap<String, Object>();
+		// If not otherwise specified last updated since the beggining of times
+		this.solutionsSelectorDefaults.put(Solution.Fields.modified, new Long(1));
+
+		//this selector forces certain criteria such as no client can impose/submit them
 		this.solutionsSelector = new HashMap<String, Object>();
 		// Fetch only active solutions
 		this.solutionsSelector.put(Solution.Fields.active, true);
 		// Fetch only Public models
 		this.solutionsSelector.put(Solution.Fields.accessTypeCode, AccessTypeCode.PB.toString());
-		// Fetch solutions last updated since the beggining of times
-		this.solutionsSelector.put(Solution.Fields.modified, new Long(1));
 
 		this.solutionRevisionsSelector = new HashMap<String, Object>();
 		// Fetch only for Public revisions
@@ -79,6 +83,15 @@ public class CatalogServiceConfiguration {
 		this.solutionsSelector = Collections.unmodifiableMap(this.solutionsSelector);;
 		this.solutionRevisionsSelector = Collections.unmodifiableMap(this.solutionRevisionsSelector);;
 	}
+
+	public Map<String, Object> getSolutionsSelectorDefaults() {
+		return this.solutionsSelectorDefaults;
+	}
+
+	public void setSolutionsSelectorDefaults(String theSelector) {
+		this.solutionsSelectorDefaults = Collections.unmodifiableMap(
+															JsonParserFactory.getJsonParser().parseMap(theSelector));
+  }
 
 	public Map<String, Object> getSolutionsSelector() {
 		return this.solutionsSelector;